BREAKING CHANGES: rename ng-functions, add cheking imports and  duplicates in ng-functions
