--- libs/ng-morph/ng/module/add-declaration-to-module.ts
@@ -1,14 +1,0 @@
-import { ClassDeclaration } from 'ts-morph';
-import { pushToArrayProperty } from '../helpers/push-to-array-property';
-
-export function addDeclarationToModule(
-  classDeclaration: ClassDeclaration,
-  declaration: string
-) {
-  pushToArrayProperty(
-    classDeclaration,
-    'NgModule',
-    'declarations',
-    declaration
-  );
-}
