--- libs/ng-morph/ng/module/add-entry-component-to-module.ts
@@ -1,14 +1,0 @@
-import { ClassDeclaration } from 'ts-morph';
-import { pushToArrayProperty } from '../helpers/push-to-array-property';
-
-export function addEntryComponentToModule(
-  classDeclaration: ClassDeclaration,
-  component: string
-) {
-  pushToArrayProperty(
-    classDeclaration,
-    'NgModule',
-    'entryComponents',
-    component
-  );
-}
