+++ libs/ng-morph/ng/module/add-entry-component-to-ng-module.ts
@@ -1,0 +1,24 @@
+import { ClassDeclaration } from 'ts-morph';
+import { checkImportAndAdd } from '../helpers/check-and-add-import';
+import { pushToArrayProperty } from '../helpers/push-to-array-property';
+
+export function addEntryComponentToNgModule(
+  classDeclaration: ClassDeclaration,
+  component: string,
+  packageName?: string
+) {
+  if (packageName) {
+    checkImportAndAdd(
+      classDeclaration.getSourceFile().getFilePath(),
+      component,
+      packageName
+    );
+  }
+
+  pushToArrayProperty(
+    classDeclaration,
+    'NgModule',
+    'entryComponents',
+    component
+  );
+}
