--- libs/ng-morph/ng/module/add-import-to-module.ts
@@ -1,9 +1,0 @@
-import { ClassDeclaration } from 'ts-morph';
-import { pushToArrayProperty } from '../helpers/push-to-array-property';
-
-export function addImportToModule(
-  classDeclaration: ClassDeclaration,
-  moduleName: string
-) {
-  pushToArrayProperty(classDeclaration, 'NgModule', 'imports', moduleName);
-}
