--- libs/ng-morph/ng/module/add-provider-to-module.ts
@@ -1,9 +1,0 @@
-import { ClassDeclaration } from 'ts-morph';
-import { pushToArrayProperty } from '../helpers/push-to-array-property';
-
-export function addProviderToModule(
-  classDeclaration: ClassDeclaration,
-  provider: string
-) {
-  pushToArrayProperty(classDeclaration, 'NgModule', 'providers', provider);
-}
