--- libs/ng-morph/ng/module/add-schema-to-module.ts
@@ -1,9 +1,0 @@
-import { ClassDeclaration } from 'ts-morph';
-import { pushToArrayProperty } from '../helpers/push-to-array-property';
-
-export function addSchemaToModule(
-  classDeclaration: ClassDeclaration,
-  schema: string
-) {
-  pushToArrayProperty(classDeclaration, 'NgModule', 'schemas', schema);
-}
