+++ libs/ng-morph/ng/module/index.ts
@@ -1,7 +1,7 @@
-export * from './add-bootstrap-to-module';
-export * from './add-declaration-to-module';
-export * from './add-entry-component-to-module';
-export * from './add-export-to-module';
-export * from './add-import-to-module';
-export * from './add-provider-to-module';
-export * from './add-schema-to-module';
+export * from './add-bootstrap-to-ng-module';
+export * from './add-declaration-to-ng-module';
+export * from './add-entry-component-to-ng-module';
+export * from './add-export-to-ng-module';
+export * from './add-import-to-ng-module';
+export * from './add-provider-to-ng-module';
+export * from './add-schema-to-ng-module';
